feat: 홈 - 진행 중인 토너먼트 api 연결
