+++ apps/web/src/consts/tournament.ts
@@ -1,0 +1,5 @@
+export const TOURNAMENT_STATUS = {
+  PENDING: 'PENDING',
+  IN_PROGRESS: 'IN_PROGRESS',
+  COMPLETED: 'COMPLETED',
+} as const;
