+++ apps/web/src/types/tournament.ts
@@ -1,1 +1,3 @@
-export type TournamentStatusT = 'PENDING' | 'IN_PROGRESS' | 'COMPLETED';
+import type { TOURNAMENT_STATUS } from '@/consts/tournament';
+
+export type TournamentStatusT = (typeof TOURNAMENT_STATUS)[keyof typeof TOURNAMENT_STATUS];
